+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Requires `hoist-core >= 21.1`
 
 ### ⚙️ Technical
+
+* Updated Admin Console's Cluster tab to refresh more frequently.
 * Alert Banner is now updated via the shared environment polling for improved simplicity and frequency.
 
 ## 67.0.0 - 2024-09-03
@@ -47,7 +49,6 @@
 * Fixed inline `SelectEditor` to ensure new value is flushed before grid editing stops.
 * `WebSocketService` now attempts to establish a new connection when app's server instance changes.
 
-
 ### ✨ Styles
 
 * Added CSS variables to support customization of `Badge` component styling.
@@ -55,7 +56,6 @@
 ### 📚 Libraries
 
 * short-unique-id `added @ 5.2`
-* hoist-core 21.0`
 
 ## 66.1.1 - 2024-08-01